requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.12.3
certifi==2025.8.3
charset-normalizer==3.4.3
colorlog==6.8.2
distro==1.9.0
feedparser==6.0.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
lxml==5.1.0
oauthlib==3.3.1
openai==2.0.1
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
tqdm==4.67.1
tweepy==4.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
